ndicom_server/requirements.txt
Django==2.0.6
djangorestframework==3.8.2
jsonfield==2.0.2
numpy==1.14.3
opencv-python==3.4.1.15
Pillow==5.1.0
psycopg2-binary==2.7.4
pytz==2018.4
scikit-learn==0.19.1
tornado==5.0.2
